backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
alembic==1.16.4
psycopg[binary]==3.2.13
pydantic-settings==2.10.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
httpx==0.28.1
celery==5.5.3
redis==6.2.0
